fix(testing-library): rollup dts files
